Add functionality to get all products from database, and get products from a category.
